+++ kotlin-source/src/main/resources/testWeb/js/rollovertd-module.js
@@ -259,7 +259,7 @@ app.controller('RolloverTDAppController', function($http, $location, $uibModal) 
                 i++;
             }
         }
-        return parseInt(value.substring(0,endIndex));
+        return parseFloat(value.substring(0,endIndex));
     }
